Added Test Field for Game.
Basic elements of this commit include:

Assets folder with Scripts, Scenes and Prefabs

Prefabs: Include the test level (for testing any new features to be added to the game. This level can be modified as need be). As well as a Prefab for a robot enemy to be included in game. (Robot Enemy currently in testing for AI behavior and movement.)

Scripts: Include a Camera Controller that follows a target player around. Player Controller (later to be replaced and updated for more sophisticated player controller down the line.) and an Enemy Behaviour script (which is the same as the hw4 AI for now im currently working on it so that it has a proper field of view).

The Scene is just the testing level.