Stage 3
Stage 3 has the following element.

There are 3 switches that need to be turned on to activate a bridge that leads to the end of the game.

To accomplish this, crates have been placed around the level and are to be used in order to turn on the switches.

Other switches are used to activate giant fans that are used to both move crates and the player as well throughout the stage3

Everything should be working right now in terms of connecting the rest of the game to stage 3.

Only need to add the third and have it be able to push the crates rather than allowing any of the pigs to push the crates. If this can not be accomplish in time that is fine as we dont necesseraliy need it.